Add self-signed cert support
